Modify netmiko_drvr:config()
Since BaseConnection defines `commit` and `save_config` all child
classes have these methods which are also callable. This means that
save_config would never actually be ran on device classes that
implemented it.

Child classes that dont implement it in netmiko throw either a
`NotImplementedError` or `AttributeError`. Switching this code to use a
try/catch so it can properly run on devices that support it.