Add "assigned_worker" to task responses.  This currently assumes only a single worker.  In the event there's more than one, whichever one started the task most recently will overwrite the other.
In the event we implement retries this would end up being the last retry, and since that'd be either the one that worked, or the last one that failed before we gave up, that's probably fine.

In any case this is set at the very beginning of job processing, so you'll see an "assigned worker" if you check the status any time after it's pulled off the queue.