Update netmiko_drvr.py
Add the ability to pass an array to "ttp_template" key in netmiko args. It will match any given command in order with the template. If there's a command without a template, it should be put in the end of the command array, and it will get the raw data.

Example 1, will match command with template, in order (command1 -> template1, command2 -> template2)
command: [ 'command1', 'command2' ]
ttp_template: [ 'template1', 'template2' ]

Example 2, will match command with template, and command without template will return the raw data (command1 -> template1, command2 -> raw)
command: [ 'command1', 'command2' ]
ttp_template: [ 'template1' ]

If command and ttp_template are a string, they will be encapsulated in an array and will be matched as before.